Ensure web app linting and API helper alignment
